requirements.txt
flask_bootstrap
flask-wtf>=0.12
wtforms
Flask>=0.10.1
GoogleAppEngineCloudStorageClient>=1.9.22.1
inflect>=0.2.5
sendgrid>=1.5.14
user-agents>=1.0.1